+++ pyramid_duh/params.py
@@ -6,6 +6,7 @@ import inspect
 import json
 from pyramid.httpexceptions import HTTPBadRequest, HTTPException
 from pyramid.interfaces import IRequest
+from pyramid.path import DottedNameResolver
 from pyramid.settings import asbool
 # pylint: disable=F0401,E0611
 from zope.interface.exceptions import DoesNotImplement
@@ -15,6 +16,7 @@ from .compat import string_type, bytes_types, num_types, is_string, is_num
 
 
 NO_ARG = object()
+__resolver__ = DottedNameResolver(__name__)
 
 
 def _param(request, name, default=NO_ARG, type=None, validate=None):
@@ -41,7 +43,8 @@ def _param(request, name, default=NO_ARG, type=None, validate=None):
 
     """
     params, loads = _params_from_request(request)
-    return _param_from_dict(request, params, name, default, type, validate, loads)
+    return _param_from_dict(request, params, name, default, type, validate,
+                            loads)
 
 
 def _params_from_request(request):
@@ -92,6 +95,16 @@ def _param_from_dict(request, params, name, default=NO_ARG, type=None,
         If the parameter is missing and no default specified
 
     """
+    type = __resolver__.maybe_resolve(type)
+    # If the type arg is wrapped with @argify, then it is a multi-param
+    # argument and retrieves its parameters directly
+    if type is not None:
+        if getattr(type, '__argify__', False):
+            return type(request.context, request)
+        elif hasattr(type, '__from_json__'):
+            if getattr(type.__from_json__, '__argify__', False):
+                return type.__from_json__(request.context, request)
+
     try:
         arg = params[name]
     except KeyError:
@@ -162,24 +175,30 @@ def argify(*args, **type_kwargs):
 
     Notes
     -----
-    Here is a sample use case::
+    Here is a sample use case:
+
+    .. code-block:: python
 
         @argify(foo=dict, ts=datetime)
         def handle_request(request, foo, ts, bar='baz'):
             # do request handling
 
-    No special type is required for strings::
+    No special type is required for strings:
+
+    .. code-block:: python
 
         @argify
         def handle_request(request, foo, bar='baz'):
             # do request handling (both 'foo' and 'bar' are strings)
 
-    If any positional arguments are missing, it will raise a HTTPBadRequest
-    exception. If any keyword arguments are missing, it will simply use
-    whatever the default value is.
+    If any positional arguments are missing, it will raise a
+    :class:`~pyramid.httpexceptions.HTTPBadRequest` exception. If any keyword
+    arguments are missing, it will simply use whatever the default value is.
 
     Note that unit tests should be unaffected by this decorator. This should be
-    valid::
+    valid:
+
+    .. code-block:: python
 
         @argify
         def myrequest(request, var1, var2='foo'):
@@ -210,51 +229,79 @@ def argify(*args, **type_kwargs):
         @functools.wraps(fxn)
         def param_twiddler(*args, **kwargs):
             """ The actual wrapper function that pulls out the params """
-            # pyramid always calls with (context, request) arguments
-            if len(args) == 2 and len(kwargs) == 0 and is_request(args[1]):
-                context, request = args[0], args[1]
-                scope = {}
+            scope = {}
+            self = None
+            # If @argify is decorating a classmethod, inject the 'cls' arg
+            # with no modification
+            if 'cls' in required:
+                args = list(args)
+                scope['cls'] = args.pop(0)
+                required.remove('cls')
 
-                params, loads = _params_from_request(request)
-                params = dict(params)
-                for param in required:
-                    type_spec = type_kwargs.get(param)
-                    if (isinstance(type_spec, tuple) or
-                            isinstance(type_spec, list)):
-                        type_def, validate = type_spec
-                    else:
-                        type_def = type_spec
-                        validate = None
-                    if param == 'context':
-                        scope['context'] = context
-                    elif param == 'request':
-                        scope['request'] = request
-                    else:
-                        scope[param] = _param_from_dict(
-                            request, params, param, NO_ARG,
-                            type_def, validate, loads=loads)
-                        params.pop(param)
-                no_val = object()
-                for param in optional:
-                    type_spec = type_kwargs.get(param)
-                    if (isinstance(type_spec, tuple) or
-                            isinstance(type_spec, list)):
-                        type_def, validate = type_spec
-                    else:
-                        type_def = type_spec
-                        validate = None
-                    val = _param_from_dict(request, params, param, no_val,
-                                           type_def, validate, loads)
-                    params.pop(param, None)
-                    if val is not no_val:
-                        scope[param] = val
-                if argspec.keywords is not None:
-                    scope.update(params)
-                return fxn(**scope)
-            else:
-                # Otherwise, it's likely a unit test. Don't alter args at all.
+            if 'self' in required:
+                self = args[0]
+                if not hasattr(self, 'request'):
+                    raise AttributeError("View class %s has no attribute "
+                                         "'request'" % self)
+                request = self.request
+                context = getattr(self, 'context', None)
+                # Multiple args passed in, it's likely a unit test.
+                # Don't alter args at all.
+                if len(args) != 1 or len(kwargs) != 0:
+                    return fxn(*args, **kwargs)
+
+            # pyramid always calls with (context, request) arguments
+            # If it doesn't, it's likely a unit test. Don't alter args at all.
+            elif not (len(args) == 2 and len(kwargs) == 0 and
+                      is_request(args[1])):
                 return fxn(*args, **kwargs)
+            else:
+                context, request = args[0], args[1]
+
+            params, loads = _params_from_request(request)
+            params = dict(params)
+            for param in required:
+                type_spec = type_kwargs.get(param)
+                if (isinstance(type_spec, tuple) or
+                        isinstance(type_spec, list)):
+                    type_def, validate = type_spec
+                else:
+                    type_def = type_spec
+                    validate = None
+
+                if param == 'context':
+                    scope['context'] = context
+                elif param == 'request':
+                    scope['request'] = request
+                elif param == 'self':
+                    scope['self'] = self
+                else:
+                    scope[param] = _param_from_dict(
+                        request, params, param, NO_ARG,
+                        type_def, validate, loads=loads)
+                    params.pop(param, None)
+            no_val = object()
+            for param in optional:
+                type_spec = type_kwargs.get(param)
+                if (isinstance(type_spec, tuple) or
+                        isinstance(type_spec, list)):
+                    type_def, validate = type_spec
+                else:
+                    type_def = type_spec
+                    validate = None
+                val = _param_from_dict(request, params, param, no_val,
+                                       type_def, validate, loads)
+                params.pop(param, None)
+                if val is not no_val:
+                    scope[param] = val
+            if argspec.keywords is not None:
+                scope.update(params)
+            return fxn(**scope)
+
+        param_twiddler.__argify__ = True
         return param_twiddler
+
+    wrapper.__argify__ = True
 
     if len(args) == 1 and len(type_kwargs) == 0 and inspect.isfunction(args[0]):
         # @argify